create current daily puzzle module
